+++ src/modelDAO/ProdutoCopaDAO.java
@@ -24,7 +24,7 @@ public class ProdutoCopaDAO implements InterfaceDAO<ProdutoCopa>{
         try {
             pstm = conexao.prepareStatement(sqlInstrucao);
             pstm.setString(1, objeto.getDescricao());
-            pstm.setFloat(2, objeto.getValor());
+            pstm.setDouble(2, objeto.getValor());
             pstm.setString(3, objeto.getObs());
             pstm.setString(4, String.valueOf(objeto.getStatus()));
             pstm.setInt(5, objeto.getCopaQuarto().getId());
@@ -127,7 +127,7 @@ public class ProdutoCopaDAO implements InterfaceDAO<ProdutoCopa>{
         try {
             pstm = conexao.prepareStatement(sqlInstrucao);
             pstm.setString(1, objeto.getDescricao());
-            pstm.setFloat(2, objeto.getValor());
+            pstm.setDouble(2, objeto.getValor());
             pstm.setString(3, objeto.getObs());
             pstm.setString(4, objeto.getStatus()+"");
             pstm.setInt(5, objeto.getCopaQuarto().getId());
